# Conflicts:
#   app/src/main/java/cookbook/frontend/fe_controllers/PopupWeeklyViewController.java
#   app/src/main/java/cookbook/frontend/fe_controllers/WeeklyViewController.java

+++ app/src/main/java/cookbook/backend/be_controllers/CommentController.java
@@ -37,7 +37,7 @@ public class CommentController {
     try (Connection conn = dbManager.getConnection(); 
       PreparedStatement pstmt = conn.prepareStatement(sql)) {
       pstmt.setString(1, commentObj.getRecipeId());
-      pstmt.setInt(2, commentObj.getUserId());
+      pstmt.setLong(2, commentObj.getUserId());
       pstmt.setString(3, commentObj.getText());
       pstmt.setString(4,commentObj.getTimestamp());
